Created new TopBar component and refactored App
