+++ src/components/About/About.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export class About extends React.Component {
+    render(): JSX.Element {
+        return <div>About</div>;
+    }
+}
